+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased 
 - Changed option `strong_branching_approx_time_limit` to `strong_branching_time_limit`
   - Default is still 100s
+- Bugfix: No bounds on binary variables in Feasibility pump [#143](https://github.com/lanl-ansi/Juniper.jl/issues/130)
 
 ### v0.4.2
 - Ability to not accept `almost` MOI solver status codes in tree search